chore(socket): updating the socket handling
