+++ src/public/game/create-game.html
@@ -109,6 +109,6 @@
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
   <script src="/socket.io/socket.io.js"></script>
-  <script src="/js/create-game.js"></script>
+  <script src="/js/game/create-game.js"></script>
 </body>
 </html>
