+++ src/public/game/lobby.html
@@ -122,7 +122,7 @@
   </div>
 
   <script src="/socket.io/socket.io.js"></script>
-  <script src="/js/lobby.js"></script>
+  <script src="/js/game/lobby.js"></script>
 </body>
 
 </html>
