+++ src/public/game/play.html
@@ -150,8 +150,7 @@
     </div>
 
     <script src="/socket.io/socket.io.js"></script>
-    <script src="/js/play.js"></script>
-    <script src="/js/chat.js"></script>
+    <script type="module" src="/js/game/play.js"></script>
 </body>
 
 </html>
