+++ src/public/game/voting-round.html
@@ -34,7 +34,7 @@
     </div>
     
     <script src="/socket.io/socket.io.js"></script>
-    <script src="js/voting-round.js"></script>
+    <script src="/js/game/voting-round.js"></script>
 </body>
 
 </html>
