+++ src/public/js/game/lobby.js
